requirements.txt
click==7.1.2
einops==0.8.0
einops_exts==0.0.4
huggingface_hub==0.23.5
ipython==8.12.3
librosa==0.10.2.post1
matplotlib==3.9.2
monotonic_align==1.2
munch==4.0.0
numpy==2.1.1
pandas==1.5.3
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.14.1
SoundFile==0.10.3.post1
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
